Speed program
I created a program that made it so that the robot could either move fast when you push the right trigger or make it slower from the base speed when you push the left trigger. I also created a simple turning mechanism that will allow the robot to either glide across the map or just simply rotate itself.